Fix mem leak

+++ inc/gfx.h
@@ -17,7 +17,7 @@ namespace gfx
     void drawText(const std::string& str, const int& x, const int& y, const uint32_t& clr);
     void drawU16Text(const std::u16string& str, const int& x, const int& y, const uint32_t& clr);
     void drawU32Text(const std::u32string& str, const int& x, const int& y, const uint32_t& clr);
-    uint32_t getTextWidth(const std::string& str);
+    size_t getTextWidth(const std::string& str);
 }
 
 #endif // GFX_H
